Refactored and moved icon methods to iconUtils file
