+++ packages/vsce/src/trees/CICSCombinedTrees/CICSCombinedLibraryTree.ts
@@ -18,7 +18,7 @@ import { CICSRegionTree } from "../CICSRegionTree";
 import { CICSTree } from "../CICSTree";
 import { CICSLibraryTreeItem } from "../treeItems/CICSLibraryTreeItem";
 import { TextTreeItem } from "../treeItems/utils/TextTreeItem";
-import { getIconOpen } from "../../utils/profileUtils";
+import { getIconOpen } from "../../utils/iconUtils";
 import { ViewMore } from "../treeItems/utils/ViewMore";
 
 export class CICSCombinedLibraryTree extends TreeItem {
