+++ packages/vsce/src/trees/CICSCombinedTrees/CICSCombinedProgramTree.ts
@@ -19,7 +19,7 @@ import { CICSRegionTree } from "../CICSRegionTree";
 import { CICSTree } from "../CICSTree";
 import { CICSProgramTreeItem } from "../treeItems/CICSProgramTreeItem";
 import { TextTreeItem } from "../treeItems/utils/TextTreeItem";
-import { getIconOpen, getIconPathInResources } from "../../utils/profileUtils";
+import { getIconOpen, getIconPathInResources } from "../../utils/iconUtils";
 import { imperative } from "@zowe/zowe-explorer-api";
 import { ViewMore } from "../treeItems/utils/ViewMore";
