+++ packages/vsce/src/trees/CICSProgramTree.ts
@@ -14,7 +14,7 @@ import { CICSProgramTreeItem } from "./treeItems/CICSProgramTreeItem";
 import { CICSRegionTree } from "./CICSRegionTree";
 import { getResource } from "@zowe/cics-for-zowe-sdk";
 import { getDefaultProgramFilter, toEscapedCriteriaString } from "../utils/filterUtils";
-import { getIconOpen } from "../utils/profileUtils";
+import { getIconOpen } from "../utils/iconUtils";
 import { toArray } from "../utils/commandUtils";
 
 export class CICSProgramTree extends TreeItem {
