+++ packages/vsce/src/trees/treeItems/web/CICSTCPIPServiceTree.ts
@@ -14,7 +14,7 @@ import { CICSTCPIPServiceTreeItem } from "./treeItems/CICSTCPIPServiceTreeItem";
 import { CICSRegionTree } from "../../CICSRegionTree";
 import { getResource } from "@zowe/cics-for-zowe-sdk";
 import { toEscapedCriteriaString } from "../../../utils/filterUtils";
-import { getIconOpen } from "../../../utils/profileUtils";
+import { getIconOpen } from "../../../utils/iconUtils";
 import { toArray } from "../../../utils/commandUtils";
 
 export class CICSTCPIPServiceTree extends TreeItem {
