+++ packages/vsce/src/utils/profileUtils.ts
@@ -10,7 +10,6 @@
  */
 
 import { ZoweVsCodeExtension, imperative } from "@zowe/zowe-explorer-api";
-import { join } from "path";
 import { window } from "vscode";
 import { ProfileManagement } from "./profileManagement";
 
@@ -25,14 +24,7 @@ export function missingSessionParameters(profileProfile: any): (string | undefin
   return missing;
 }
 
-export function getIconPathInResources(iconFileNameLight: string, iconFileNameDark: string): { light: string; dark: string } {
-  return {
-    // We bundle the extension into a single `dist/extension.js`
-    // `__dirname/../resources/imgs === `/path/to/dist/../resources/imgs`
-    light: join(__dirname, "..", "resources", "imgs", iconFileNameLight),
-    dark: join(__dirname, "..", "resources", "imgs", iconFileNameDark),
-  };
-}
+
 
 export async function promptCredentials(sessionName: string, rePrompt?: boolean): Promise<imperative.IProfileLoaded> {
   // const mProfileInfo = new ProfileInfo("zowe", {
@@ -48,37 +40,4 @@ export async function promptCredentials(sessionName: string, rePrompt?: boolean)
     window.showInformationMessage("Input credentials operation Cancelled");
   }
   return promptInfo;
-}
-
-export function getIconOpen(open: boolean = true) {
-  return getIconPathInResources(
-    `folder-${open ? "open" : "closed"}-dark.svg`,
-    `folder-${open ? "open" : "closed"}-light.svg`);
-}
-
-export function getIconByStatus(resourceType: string, resourceTreeItem: any) {
-  switch (resourceType) {
-    case "PROGRAM":
-      return resourceTreeItem.status === "DISABLED"
-        ? getIconPathInResources("program-disabled-dark.svg", "program-disabled-light.svg")
-        : getIconPathInResources("program-dark.svg", "program-light.svg");
-    case "TRANSACTION":
-      return resourceTreeItem.status === "DISABLED"
-        ? getIconPathInResources("local-transaction-disabled-dark.svg", "local-transaction-disabled-light.svg")
-        : getIconPathInResources("local-transaction-dark.svg", "local-transaction-light.svg");
-    case "LOCAL_FILE":
-      return resourceTreeItem.openstatus === "CLOSED" && resourceTreeItem.enablestatus === "DISABLED"
-        ? getIconPathInResources("local-file-disabled-closed-dark.svg", "local-file-disabled-closed-light.svg")
-        : resourceTreeItem.openstatus === "CLOSED"
-          ? getIconPathInResources("local-file-closed-dark.svg", "local-file-closed-light.svg")
-          : resourceTreeItem.enablestatus === "DISABLED"
-            ? getIconPathInResources("local-file-disabled-dark.svg", "local-file-disabled-light.svg")
-            : getIconPathInResources("local-file-dark.svg", "local-file-light.svg");
-    case "TASK":
-      return resourceTreeItem.runstatus === "RUNNING"
-        ? getIconPathInResources("task-running-dark.svg", "task-running-light.svg")
-        : resourceTreeItem.runstatus === "SUSPENDED"
-          ? getIconPathInResources("task-suspended-dark.svg", "task-suspended-light.svg")
-          : getIconPathInResources("task-dispatched-dark.svg", "task-dispatched-light.svg");
-  }
 }
